requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-restful==0.3.10
gunicorn==21.2.0
gevent==23.9.1
sqlalchemy==2.0.23
chromadb==0.4.22
langchain==0.1.0
langchain-community==0.0.10
paddleocr==2.7.3
paddlepaddle==2.6.2
vllm==0.2.0
transformers==4.36.0
accelerate==0.25.0
bitsandbytes==0.41.0
httpx==0.25.0
python-multipart==0.0.6
pydantic==1.10.13
python-dotenv==1.0.0
gitpython==3.1.43
